Add files via upload
The cost of a stock on each day is given in an array, find the max profit and print the index ,that you can make by buying and selling in those days. It is an effective approach to find the profit and indexes with time complexity O(n) and Auxiliary space O(1).